+++ docs/projects/projects.mdx
@@ -1,7 +1,7 @@
 ---
 title: Projects
 ---
-import Expandable from '@site/src/components/Expandable';
+import Expandable from '@site/src/components/Expandable/Expandable';
 
 ## My Projects
